[Bugfix] Fix Qwen3-TTS Base voice cloning: fp32 encoder + chunk boundary
Two bugs preventing Base (voice-clone) task from producing correct audio:

1. Speech tokenizer encoder ran in bfloat16, causing ~50% of encoded
   reference-audio codes to diverge from float32 baseline. The corrupted
   prompt prevents the talker from generating stop token 2150, producing
   ~318s of audio instead of ~8s. Fix: load encoder in float32.

2. Cherry-pick chunk boundary fix (#2378): off-by-one in initial chunk
   phase boundary check caused the final codec chunk to be malformed
   (length 1, not divisible by 16 quantizers), resulting in 0-byte output
   even when stop token was correctly generated.

+++ vllm_omni/model_executor/models/qwen3_tts/qwen3_tts_talker.py
@@ -1116,9 +1116,13 @@ class Qwen3TTSTalkerForConditionalGeneration(nn.Module):
         if preprocessor_config_path is None:
             raise ValueError(f"{self.model_path}/speech_tokenizer/preprocessor_config.json not found")
         speech_tokenizer_dir = os.path.dirname(speech_tokenizer_path)
+        # Use float32 for the speech tokenizer encoder to avoid bfloat16
+        # precision loss in VQ codebook lookups.  bf16 causes ~50% of encoded
+        # reference-audio codes to differ from the float32 baseline, which
+        # corrupts the talker prompt and prevents stop-token generation.
         tok = Qwen3TTSTokenizer.from_pretrained(
             speech_tokenizer_dir,
-            torch_dtype=torch.bfloat16,
+            torch_dtype=torch.float32,
         )
         # Prefer GPU for encoder if available; otherwise keep CPU.
         dev = next(self.parameters()).device
